requirements.txt
anyio==3.6.2
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1640817743617/work
argon2-cffi-bindings @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi-bindings_1666850768662/work
autopep8 @ file:///home/conda/feedstock_root/build_artifacts/autopep8_1677845287175/work
bcrypt @ file:///home/conda/feedstock_root/build_artifacts/bcrypt_1666831240012/work
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1671179353105/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
cryptography==40.0.1
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1673812078720/work
ecdsa @ file:///home/conda/feedstock_root/build_artifacts/ecdsa_1657437522792/work
email-validator==1.3.1
fastapi==0.92.0
flit_core @ file:///home/conda/feedstock_root/build_artifacts/flit-core_1684084314667/work/source/flit_core
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1666808654411/work
greenlet @ file:///home/conda/feedstock_root/build_artifacts/greenlet_1684082927543/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.31
orjson==3.8.9
passlib==1.7.4
peewee @ file:///home/conda/feedstock_root/build_artifacts/peewee_1682115678636/work
protobuf==4.21.12
pyasn1==0.4.8
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1669306857274/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==1.10.5
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1683676063469/work
PyMySQL==1.0.2
python-dotenv==1.0.0
python-jose @ file:///home/conda/feedstock_root/build_artifacts/python-jose_1624701740191/work
python-multipart==0.0.6
PyYAML==6.0
rfc3986==1.5.0
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1658328885051/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/sqlalchemy_1684569061937/work
starlette==0.25.0
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1644342247877/work
typing==3.7.4.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685045523825/work
ujson==5.7.0
uvicorn @ file:///home/conda/feedstock_root/build_artifacts/uvicorn-split_1682667763783/work
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
